requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
Authlib==0.11
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.4
Pillow==6.0.0
pony==0.7.10
protobuf==3.8.0
pycparser==2.19
PyYAML==5.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
gunicorn==19.9.0
flask-cors